Refactor class locations
