+++ src/Models/PersonalAccessToken.php
@@ -3,10 +3,9 @@
 namespace Cratespace\Sentinel\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use Cratespace\Sentinel\Contracts\Auth\Access;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 
-class PersonalAccessToken extends Model implements Access
+class PersonalAccessToken extends Model
 {
     /**
      * The attributes that should be cast to native types.
@@ -66,30 +65,5 @@ class PersonalAccessToken extends Model implements Access
         if ($instance = static::find($id)) {
             return hash_equals($instance->token, hash('sha256', $token)) ? $instance : null;
         }
-    }
-
-    /**
-     * Determine if the token has a given ability.
-     *
-     * @param string $ability
-     *
-     * @return bool
-     */
-    public function can(string $ability): bool
-    {
-        return in_array('*', $this->abilities) ||
-            array_key_exists($ability, array_flip($this->abilities));
-    }
-
-    /**
-     * Determine if the token is missing a given ability.
-     *
-     * @param string $ability
-     *
-     * @return bool
-     */
-    public function cant(string $ability): bool
-    {
-        return ! $this->can($ability);
     }
 }
