+++ tests/HasApiTokensTest.php
@@ -5,8 +5,8 @@ namespace Cratespace\Sentinel\Tests;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Database\Eloquent\Model;
 use Cratespace\Sentinel\Auth\Tokens\TransientToken;
-use Cratespace\Sentinel\Models\PersonalAccessToken;
 use Cratespace\Sentinel\Models\Traits\HasApiTokens;
+use Cratespace\Sentinel\Auth\Tokens\PersonalAccessToken;
 
 class HasApiTokensTest extends TestCase
 {
